ALERT: the old effects.gwrp file will no longer work. You will have to create custom effects again (see below).
* Changed the setting code so that it uses less insane lists and now 
uses a matrix system (basically just flipped x and y's from the previous 
matrix). This was so that we could check for savers and movie files more 
easily (and it cleans up a few things).

* Added a bunch more presets, more to come (to get updates in the 
future, delete or rename the existing presets.gwrp file and run 'svn up' 
again [your presets will be lost in the process]).