* Fiddled with the tips and welcome messages and their formatting. * Added preferences icon to titlebar.
IMPORTANT TODO: Get rid of the stupid repetitive glade stuff: The 
warnings about uninstalled stuff should all be the same window, and the 
"Adding a new effect" and the "Getting Started" messages should be added 
dynamically depending on the situation. It's absurd to have such 
wastefulness.

Also: Should get to work on the search bar on the top of the main list. 
It's in the gladefile but invisible and not functional. Should be 
simple.